1.2.0.7 updates

+++ TranscendPlugins/InventoryEnhancements/AutoTrash.cs
@@ -7,18 +7,13 @@ namespace GTRPlugins
         public static void Trash()
         {
             Player player = Main.player[Main.myPlayer];
-            int i = 10;
-            while (i < 51)
+            for (int i = 10; i < 51; i++)
             {
-                if (InventoryEnhancements.config.TrashList.Contains(player.inventory[i].type))
+                if (Inventory_Enhancements.config.TrashList.Contains(player.inventory[i].type) && !player.inventory[i].favorited)
                 {
-                    if (!player.inventory[i].favorited)
-                    {
-                        player.trashItem = player.inventory[i];
-                        player.inventory[i] = new Item();
-                    }
+                    player.trashItem = player.inventory[i];
+                    player.inventory[i] = new Item();
                 }
-                i++;
             }
         }
     }
